Fixed BuyOrder constructor

+++ Orders/BuyOrder.cs
@@ -6,6 +6,6 @@ namespace Virtual_Trading_Simulator_Project.Orders;
 
 public class BuyOrder : Order
 {
-    public BuyOrder(Trader trader, int quantity, Ticker security, ITradeStrategy tradeStrategy)
+    public BuyOrder(Trader trader, double quantity, Ticker security, ITradeStrategy tradeStrategy)
         : base(trader, quantity, security, tradeStrategy){}
 }
